Basic path finding and store items in a list instead of a 2D array

+++ HYYBLO_prog3/House.cs
@@ -13,7 +13,7 @@ namespace HYYBLO_prog3
             Image = new System.Windows.Media.Imaging.BitmapImage(new Uri("D:/Dokumentumok/Visual Studio 2015/Projects/oenik_prog3_2017osz_hyyblo/HYYBLO_prog3/Images/Buildings/house" + n + "-" + floors + ".png"));
         }
 
-        public override Rect GenerateRect(int x, int y, int cell)
+        public override Rect GenerateRect(double x, double y, int cell)
         {
             return new Rect(x, y - (floors - 1) * (cell/4), cell, cell + (cell/4) * (floors - 1));
         }
